refactor: split PatchFactory into specialized subclasses
