+++ src/lib/index.ts
@@ -1,3 +1,2 @@
 export type { IVoxel, IVoxelMap, IVoxelMaterial } from './terrain/i-voxel-map';
 export { EPatchComputingMode, Terrain } from './terrain/terrain';
-
